Avoid having to know if features are 0- or 1-indexed
We were observing a problem where on some systems the first feature in a
layer would be feat[0], while on others it would be feat[1].  Matthew
Somerville (@dracos) worked out that this depended on which driver was
being used to load the KML files: KML (0-indexed) or libkml (1-indexed).

Rather than try to work out that detail, we can avoid the problem by
using `next` to get the first feature in the layer.