Add a command to set .country on areas (and assess ambiguities)
This command:

 - Iterates over every area

 - Takes an arbitrary point within every polygon in that area

 - Finds all the O02 (country-level) areas that contain those points

 - If there's exactly one of those, sets it as the country on the area

 - Otherwise sets it to a special '?' country, and outputs an error
   message describing the ambiguity

The plan is to run this command (*without* --commit) to assess the
number of areas that have ambiguous enclosing areas, and then decide
what do about those cases. (e.g. we could change MapIt's data model to
accommodate multiple countries for an area, making sure not to break the
existing API; we could pick the smaller of the ambiguous regions
(typically more useful information); we could use the special '?'
country for real.)